requirements.txt

psycopg2 == 2.8.6
python-dotenv == 0.14.0
pytz == 2020.1
matplotlib == 3.3.2

